backend/requirements.txt
anyio==3.6.2
black==22.6.0
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
deptry==0.6.5
distlib==0.3.6
fastapi==0.87.0
fastapi-sessions==0.3.2
fastapi-utils==0.2.1
filelock==3.8.0
flake8==4.0.1
Flask==2.1.3
h11==0.14.0
httptools==0.5.0
identify==2.5.8
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.5.1
pathspec==0.9.0
pdfminer.six==20221105
Pillow==9.3.0
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
pyflakes==2.4.0
PyMuPDF==1.21.0
PyPDF2==2.11.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
toml==0.10.2
tomli==2.0.1
torch==1.13.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
virtualenv==20.16.7
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.1
